correct

+++ src/main/java/com/netcracker/repository/BoxRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface BoxRepository extends JpaRepository<Box, Long> {
+public interface BoxRepository extends JpaRepository<Box, Integer> {
 
     List<Box> findBoxByClientId(Integer clientId);
 }
